added exceptions

+++ src/main/java/pl/jcommerce/joannajaromin/studentbook/controller/HomeworkController.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import pl.jcommerce.joannajaromin.studentbook.dto.HomeworkDtoWithoutFile;
 import pl.jcommerce.joannajaromin.studentbook.dto.SaveHomeworkDto;
+import pl.jcommerce.joannajaromin.studentbook.exception.HomeworkNotFoundException;
 import pl.jcommerce.joannajaromin.studentbook.service.HomeworkService;
 
 import javax.validation.Valid;
@@ -29,37 +30,68 @@ public class HomeworkController {
     @PostMapping(value = "/homeworks", consumes = "multipart/form-data")
     @ResponseBody
     public HomeworkDtoWithoutFile uploadHomework(@RequestPart("uploadFile") MultipartFile file,
-                                      @RequestPart("saveHomeworkDto") @Valid SaveHomeworkDto saveHomeworkDto){
-        HomeworkDtoWithoutFile homeworkDto = homeworkService.saveHomework(file, saveHomeworkDto);
-        return homeworkDto;
+                                                 @Valid @RequestPart("saveHomeworkDto") SaveHomeworkDto saveHomeworkDto) {
+        // it doesn't work. When file is not attached status is 400, but the message isn't printed
+//        if (!file.isEmpty()){
+//            HomeworkDtoWithoutFile homeworkDto = homeworkService.saveHomework(file, saveHomeworkDto);
+//            return homeworkDto;
+//        }
+//        else throw new HomeworkFileNotAttached("Brak załączonego pliku");
+        return homeworkService.saveHomework(file,saveHomeworkDto);
     }
 
     @GetMapping("/homeworks/{homeworkId}")
     public HomeworkDtoWithoutFile getHomework (@PathVariable int homeworkId){
-        HomeworkDtoWithoutFile homeworkDtoWithoutFile = homeworkService.findById(homeworkId);
-        return homeworkDtoWithoutFile;
+            HomeworkDtoWithoutFile homeworkDtoWithoutFile = homeworkService.findById(homeworkId);
+            if (homeworkDtoWithoutFile != null) {
+                return homeworkDtoWithoutFile;
+            } else {
+                throw new HomeworkNotFoundException("Brak zadania domowego o id = " + homeworkId);
+            }
+//        I tried surrounding it with try/catch, but it doesn't work. When id is not integer there is
+//        MethodArgumentTypeMismatchException thrown (status 400) without custom message
+//        there seems to be the problem with PathVariable annotations:
+//        https://github.com/spring-projects/spring-framework/issues/11041
+//        catch (MethodArgumentTypeMismatchException exc){
+//            throw new HomeworkNotFoundException("Brak zadania domowego o id = " + homeworkId);
+//        }
     }
 
     @GetMapping("/homeworks")
     public List<HomeworkDtoWithoutFile> getAllHomeworks(){
-        return homeworkService.findAll();
+        List<HomeworkDtoWithoutFile> homeworkDtoWithoutFiles = homeworkService.findAll();
+        if (homeworkDtoWithoutFiles == null){
+            throw new HomeworkNotFoundException("Brak zadań do wyświetlenia.");
+        }
+        else {
+            return homeworkService.findAll();
+        }
     }
 
     @GetMapping("/downloadHomework/{fileId}")
     public ResponseEntity<ByteArrayResource> downloadFile (@PathVariable int fileId){
-        ByteArrayResource resource = homeworkService.downloadFile(fileId);
         HomeworkDtoWithoutFile homeworkDtoWithoutFile = homeworkService.findById(fileId);
-        HttpHeaders headers = new HttpHeaders();
-        headers.add(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename="+homeworkDtoWithoutFile.getFileName());
-        return ResponseEntity.ok()
-                .headers(headers)
-                .contentLength(resource.contentLength())
-                .contentType(MediaType.parseMediaType("application/octet-stream"))
-                .body(resource);
+        if (homeworkDtoWithoutFile == null){
+            throw new HomeworkNotFoundException("Nie odnaleziono pliku o id = " + fileId);
+        }
+        else {
+            ByteArrayResource resource = homeworkService.downloadFile(fileId);
+            HttpHeaders headers = new HttpHeaders();
+            headers.add(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + homeworkDtoWithoutFile.getFileName());
+            return ResponseEntity.ok()
+                    .headers(headers)
+                    .contentLength(resource.contentLength())
+                    .contentType(MediaType.parseMediaType("application/octet-stream"))
+                    .body(resource);
+        }
     }
 
     @DeleteMapping("/homeworks/{homeworkId}")
     public void deleteHomework (@PathVariable int homeworkId){
+        HomeworkDtoWithoutFile homeworkDtoWithoutFile = homeworkService.findById(homeworkId);
+        if (homeworkDtoWithoutFile == null){
+            throw new HomeworkNotFoundException("Nie odnaleziono zadania domowego o id = " + homeworkId);
+        }
         homeworkService.deleteById(homeworkId);
     }
